feat: 각 스크랩 타입별 조회 및 스크랩 응답 반환 구현

+++ src/main/java/com/commerce/pagopa/scrap/application/dto/response/ScrapListResponseDto.java
@@ -6,7 +6,7 @@ import java.util.List;
 import java.util.Map;
 
 public record ScrapListResponseDto(
-        Map<String, Long> count,
+        Map<String, Integer> count,
         UserResponseDto user,
         List<ScrapCollectionItem> collection
 ) {
